# docs/swagger.yaml
definitions:
...
  controllers.FinishEvent:
    properties:
      anchorAddress:
        type: string
      anchorId:
        type: integer
      anchorName:
        type: string
      blockNumber:
        type: integer
      chainId:
        type: integer
      contractAddress:
        type: string
      eventType:
        type: string
      gasPrice:
        type: string
      gasUsed:
        type: string
      hash:
        type: string
      networkId:
        type: integer
      remoteChainId:
        type: integer
      remoteNetworkId:
        type: integer
      status:
        type: integer
      timestamp:
        type: integer
      tokenListKey:
        type: string
      tokenName:
        type: string
      txId:
        type: string
    type: object
  controllers.FinishEventView:
    properties:
      count:
        type: integer
      finishEventList:
        items:
          $ref: '#/definitions/controllers.FinishEvent'
        type: array
    type: object
  controllers.JsonResult:
    properties:
      code:
        type: integer
      data:
        type: object
      err_msg:
        type: string
    required:
    - code
    type: object
...
  dao.TokenListCount:
    properties:
      count:
        type: integer
      date:
        type: string
      fee:
        type: integer
      timeType:
        type: string
    type: object
...
  dao.TxAnchorsNode:
    properties:
      anchor_id:
        type: integer
      anchorAddress:
        description: 锚定节点地址
        type: string
      chain_id:
        type: integer
      contractAddress:
        description: 跨链合约地址
        type: string
      count:
        type: string
      date:
        type: string
      fee:
        type: integer
      name:
        type: string
      source_chain_id:
        type: integer
      source_network_id:
        type: integer
      target_chain_id:
        type: integer
      target_network_id:
        type: integer
      timeType:
        type: string
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /chart/anchorCount/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: hour:2020-07-10 12:00:00 day:2020-07-10 week:202025
        in: query
        name: startTime
        required: true
        type: string
      - description: hour:2020-07-12 12:00:00 day:2020-07-12 week:202029
        in: query
        name: endTime
        required: true
        type: string
      - description: 1,2
        in: query
        name: tokenKey
        required: true
        type: string
      - description: hour,day,week
        in: query
        name: timeType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dao.TokenListCount'
              type: object
      summary: 签名监控
      tags:
      - Chart
  /chart/crossMonitor/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 1,2
        in: query
        name: tokenKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/controllers.FinishEventView'
              type: object
      summary: 锚定节点监控
      tags:
      - Chart
...
  /chart/feeAndCount/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: hour:2020-07-10 12:00:00 day:2020-07-10 week:202025
        in: query
        name: startTime
        required: true
        type: string
      - description: hour:2020-07-12 12:00:00 day:2020-07-12 week:202027
        in: query
        name: endTime
        required: true
        type: string
      - description: chainId
        in: query
        name: chainId
        required: true
        type: integer
      - description: hour,day,week
        in: query
        name: timeType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.JsonResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dao.TxAnchorsNode'
                  type: array
              type: object
      summary: 统计锚定节点验证数和手续费
      tags:
      - Chart
  /chart/finishList/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 时间戳
        in: query
        name: startTime
        type: string
      - description: 时间戳
        in: query
        name: endTime
        type: string
      - description: 锚定节点ID
        in: query
        name: anchorId
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: string
      - description: 页数
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/controllers.FinishEventView'
              type: object
      summary: MakeFinish手续费记录
      tags:
      - Chart
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"